refactor: move page to features
