+++ site/src/features/pages/single_train/index.ts
@@ -1,0 +1,3 @@
+export { TrainPage } from './components/page'
+
+export type { TrainPageProps } from './components/page'
