+++ site/src/features/pages/single_train/tests/helpers.test.ts
@@ -1,0 +1,7 @@
+import { getFormattedDate } from '../helpers'
+
+import { it, expect } from 'vitest'
+
+it("doesn't throw on an invalid time", () => {
+  expect(() => getFormattedDate('test')).not.toThrow()
+})
